OverflowStack cleanup
